Stop

Stop provides the locations of stops for transit lines. More details about the physical characteristics of of each stop can be found from StopPlace The stop API provides ScheduledStopPoint which describes individual stops as well as StopAreas which describes the parent heirarchy of stops.

get_stops(operator_id)

Arguments

operator_id
ID of the transit operator to filter the stops

Value

Dataframe of ScheduledStopPlace

Examples

get_stops(operator_id='SF')

#> [1] 200

#>          id                             Name Location.Longitude
#>    1: 15829           100 O'Shaughnessy Blvd        -122.450678
#>    2: 14343                    101 Dakota St         -122.39565
#>    3: 14122              1095 CONNECTICUT ST        -122.397341
#>    4: 13230             10th Ave & Ortega St        -122.466502
#>    5: 13231            10th Ave & Pacheco St        -122.466376
#>   ---                                                          
#> 3552: 16986  Yerba Buena Ave & Brentwood Ave        -122.459189
#> 3553: 16988  Yerba Buena Ave & Hazelwood Ave        -122.459705
#> 3554: 16987  Yerba Buena Ave & Hazelwood Ave        -122.459659
#> 3555: 16989   Yerba Buena Ave & Ravenwood Dr         -122.46055
#> 3556: 16990 Yerba Buena Ave & Saint Elmo Way        -122.458719
#>       Location.Latitude    StopType
#>    1:         37.744481 onstreetBus
#>    2:          37.75374 onstreetBus
#>    3:         37.753902 onstreetBus
#>    4:         37.752724 onstreetBus
#>    5:          37.75086 onstreetBus
#>   ---                              
#> 3552:         37.733702 onstreetBus
#> 3553:         37.734389 onstreetBus
#> 3554:         37.734559 onstreetBus
#> 3555:          37.73531 onstreetBus
#> 3556:         37.732623 onstreetBus

get_stops(operator_id='AC')

#> [1] 200

#>          id                       Name Location.Longitude Location.Latitude
#>    1: 53326          104th Av:Birch St       -122.1620378        37.7416345
#>    2: 53540          104th Av:Birch St       -122.1619362        37.7415252
#>    3: 51187       104th Av:Plymouth St       -122.1644793        37.7410636
#>    4: 55206       104th Av:Plymouth St       -122.1646511        37.7408899
#>    5: 59151      104th Av:Sunnyside St       -122.1596029        37.7421904
#>   ---                                                                      
#> 5356: 50558      W Winton Av:Saklan Rd       -122.1177442        37.6530046
#> 5357: 50880 W Winton Av:Santa Clara St       -122.0980041        37.6578801
#> 5358: 56866 W Winton Av:Santa Clara St        -122.097748        37.6577531
#> 5359: 55673      W Winton Av:Wright Dr       -122.1091506        37.6533541
#> 5360: 56330      W Winton Av:Wright Dr       -122.1087171        37.6537791
#>          StopType
#>    1: onstreetBus
#>    2: onstreetBus
#>    3: onstreetBus
#>    4: onstreetBus
#>    5: onstreetBus
#>   ---            
#> 5356: onstreetBus
#> 5357: onstreetBus
#> 5358: onstreetBus
#> 5359: onstreetBus
#> 5360: onstreetBus