Retrieve a list of traffic events from today.

Retrieve a list of traffic events from today.

traffic_events(status = "ACTIVE", headline = NA, event_type = NA,
  severity = NA, updated = NA, schedule = NA, limit = 10000)

Arguments

status
Default 'ALL'.

headline
Summary of event.

event_type
Type of event such as CONSTRUCTION, SPECIAL_EVENT

severity
level of severity of event. Note this is usually UNKNOWN

updated
UTC timestamp of event updated on or after.

schedule
UTC timestamp of times event is active

limit
Limiting number of events to return

Value

Dataframe of events filtered to parameters passed

Examples

traffic_events()

#> [1] 200

#>                                url                          jurisdiction_url
#>  1: /traffic/events/511.org/105877 http://api.511.org/jurisdictions/511.org/
#>  2: /traffic/events/511.org/113553 http://api.511.org/jurisdictions/511.org/
#>  3: /traffic/events/511.org/129962 http://api.511.org/jurisdictions/511.org/
#>  4: /traffic/events/511.org/192477 http://api.511.org/jurisdictions/511.org/
#>  5: /traffic/events/511.org/196504 http://api.511.org/jurisdictions/511.org/
#>  6: /traffic/events/511.org/197674 http://api.511.org/jurisdictions/511.org/
#>  7: /traffic/events/511.org/210721 http://api.511.org/jurisdictions/511.org/
#>  8: /traffic/events/511.org/211508 http://api.511.org/jurisdictions/511.org/
#>  9: /traffic/events/511.org/216774 http://api.511.org/jurisdictions/511.org/
#> 10: /traffic/events/511.org/220809 http://api.511.org/jurisdictions/511.org/
#> 11: /traffic/events/511.org/240795 http://api.511.org/jurisdictions/511.org/
#> 12: /traffic/events/511.org/246839 http://api.511.org/jurisdictions/511.org/
#> 13: /traffic/events/511.org/247749 http://api.511.org/jurisdictions/511.org/
#> 14: /traffic/events/511.org/253470 http://api.511.org/jurisdictions/511.org/
#> 15: /traffic/events/511.org/253498 http://api.511.org/jurisdictions/511.org/
#> 16: /traffic/events/511.org/253500 http://api.511.org/jurisdictions/511.org/
#> 17: /traffic/events/511.org/253502 http://api.511.org/jurisdictions/511.org/
#> 18: /traffic/events/511.org/253507 http://api.511.org/jurisdictions/511.org/
#> 19: /traffic/events/511.org/253514 http://api.511.org/jurisdictions/511.org/
#> 20: /traffic/events/511.org/253543 http://api.511.org/jurisdictions/511.org/
#> 21: /traffic/events/511.org/253544 http://api.511.org/jurisdictions/511.org/
#> 22: /traffic/events/511.org/253545 http://api.511.org/jurisdictions/511.org/
#> 23: /traffic/events/511.org/253590 http://api.511.org/jurisdictions/511.org/
#> 24: /traffic/events/511.org/253591 http://api.511.org/jurisdictions/511.org/
#> 25: /traffic/events/511.org/253596 http://api.511.org/jurisdictions/511.org/
#> 26: /traffic/events/511.org/253597 http://api.511.org/jurisdictions/511.org/
#> 27: /traffic/events/511.org/253599 http://api.511.org/jurisdictions/511.org/
#> 28: /traffic/events/511.org/253600 http://api.511.org/jurisdictions/511.org/
#> 29: /traffic/events/511.org/253601 http://api.511.org/jurisdictions/511.org/
#> 30: /traffic/events/511.org/253602 http://api.511.org/jurisdictions/511.org/
#> 31: /traffic/events/511.org/253603 http://api.511.org/jurisdictions/511.org/
#>                                url                          jurisdiction_url
#>                 id status
#>  1: 511.org/105877 ACTIVE
#>  2: 511.org/113553 ACTIVE
#>  3: 511.org/129962 ACTIVE
#>  4: 511.org/192477 ACTIVE
#>  5: 511.org/196504 ACTIVE
#>  6: 511.org/197674 ACTIVE
#>  7: 511.org/210721 ACTIVE
#>  8: 511.org/211508 ACTIVE
#>  9: 511.org/216774 ACTIVE
#> 10: 511.org/220809 ACTIVE
#> 11: 511.org/240795 ACTIVE
#> 12: 511.org/246839 ACTIVE
#> 13: 511.org/247749 ACTIVE
#> 14: 511.org/253470 ACTIVE
#> 15: 511.org/253498 ACTIVE
#> 16: 511.org/253500 ACTIVE
#> 17: 511.org/253502 ACTIVE
#> 18: 511.org/253507 ACTIVE
#> 19: 511.org/253514 ACTIVE
#> 20: 511.org/253543 ACTIVE
#> 21: 511.org/253544 ACTIVE
#> 22: 511.org/253545 ACTIVE
#> 23: 511.org/253590 ACTIVE
#> 24: 511.org/253591 ACTIVE
#> 25: 511.org/253596 ACTIVE
#> 26: 511.org/253597 ACTIVE
#> 27: 511.org/253599 ACTIVE
#> 28: 511.org/253600 ACTIVE
#> 29: 511.org/253601 ACTIVE
#> 30: 511.org/253602 ACTIVE
#> 31: 511.org/253603 ACTIVE
#>                 id status
#>                                                                                                                                                                   headline
#>  1:        Caltrans : Long-term construction on Van Ness Ave Northbound between Market St (San Francisco) and Lombard St (San Francisco). Left lane closed. Expect delays.
#>  2: Caltrans : Emergency road closure on CA-35 Northbound and Southbound between CA-9 (Los Gatos) and Black Rd (Los Gatos). Lanes closed. Consider using alternate routes.
#>  3:                                                      Caltrans : Long-term construction on CA-13 Southbound north of I-580 E (Oakland). Off Ramp closed. Expect delays.
#>  4:                                                   Caltrans : Long-term construction on I-880 Northbound exiting at 29TH Ave (Oakland). Off Ramp closed. Expect delays.
#>  5:            Caltrans : Long-term construction on CA-1 Northbound and Southbound north of Gravenstein Hwy (Valley Ford). Lane closed. One way traffic control in effect.
#>  6:             Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Gravenstein Hwy (Valley Ford). Lane closed. One way traffic control in effect.
#>  7:           Caltrans : Long-term construction on CA-1 Southbound between Muir Woods Rd (Unincorporated) and Sir Francis Drake Blvd (Olema). Lanes closed. Expect delays.
#>  8:            Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Muir Woods Rd (Unincorporated). Lane closed. One way traffic control in effect.
#>  9:                 Caltrans : Long-term construction on CA-1 Northbound and Southbound at Muir Woods Rd (Unincorporated). Lane closed. One way traffic control in effect.
#> 10:              Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Valley Ford Rd (Valley Ford). Lane closed. One way traffic control in effect.
#> 11:                                         Caltrans : Long-term construction on CA-82 Northbound at S Spruce Ave (South San Francisco). Right lane closed. Expect delays.
#> 12:        Caltrans : Long-term construction on Van Ness Ave Southbound between Lombard St (San Francisco) and Market St (San Francisco). Left lane closed. Expect delays.
#> 13:                                             Caltrans : Emergency lane closure on I-80 Westbound at Bay Bridge - Toll Plaza (Oakland). Toll lane closed. Expect delays.
#> 14:                               Caltrans : Roadwork on CA-1 Northbound and Southbound at Gravenstein Hwy (Valley Ford). Lanes closed. One way traffic control in effect.
#> 15:                                                         Caltrans : Roadwork on US-101 Northbound entering at Divisadero St (San Francisco). On Ramp closed until 2 PM.
#> 16:                                                        Caltrans : Roadwork on CA-82 Southbound at Lincoln St (Santa Clara). Right lane and shoulder closed until 3 PM.
#> 17:                                            Caltrans : Roadwork on CA-82 Northbound and Southbound at Holly St (San Carlos). Right lane and shoulder closed until 3 PM.
#> 18:                                                         Caltrans : Roadwork on US-101 Southbound exiting at Divisadero St (San Francisco). Off Ramp closed until 7 PM.
#> 19:                                  Caltrans : Roadwork on CA-128 Eastbound and Westbound at IDA Clayton Rd (Calistoga). Lanes closed. One way traffic control in effect.
#> 20:                                          Caltrans : Roadwork on I-280 Southbound between CA-92 E (San Mateo) and Edgewood Rd (Unincorporated). Lane closed until 4 PM.
#> 21:                           Caltrans : Bridge work on I-580 Westbound between Harbour Way S (Richmond) and Garrard Blvd (Richmond). Various lanes closed until midnight.
#> 22:                                                         Caltrans : Bridge work on I-580 Westbound entering at Harbour Way S (Richmond). On Ramp closed until midnight.
#> 23:                                                                    CHP : Accident on I-80 Eastbound west of Cummings Skwy (Crockett). Off Ramp blocked. Expect delays.
#> 24:                                                CHP : Disabled vehicle on I-80 Eastbound west of I-580 W-macarthur Maze (Oakland). Center lanes blocked. Expect delays.
#> 25:                                                 CHP : Obstruction on I-80 Westbound east of Treasure Island Rd (San Francisco). Center Divider blocked. Expect delays.
#> 26:                                                    CHP : Accident with injuries on US-101 Northbound west of Woodside Rd (Redwood City). Lanes blocked. Expect delays.
#> 27:                                                           CHP : Obstruction on US-101 Southbound north of E San Martin Ave (San Martin). Lanes blocked. Expect delays.
#> 28:                                                      CHP : Accident with injuries on I-280 Southbound north of Mclaughlin Ave (San Jose). Lane blocked. Expect delays.
#> 29:                                                         CHP : Accident on US-101 Southbound south of Cochrane Rd (Morgan Hill). Right shoulder blocked. Expect delays.
#> 30:                                                                CHP : Accident on US-101 Northbound south of I-880 N (San Jose). Right shoulder blocked. Expect delays.
#> 31:                                                            CHP : Obstruction on I-280 Southbound north of El Monte Rd (Los Altos Hills). Lanes blocked. Expect delays.
#>                                                                                                                                                                   headline
#>       event_type severity           created           updated geography.type
#>  1: CONSTRUCTION  UNKNOWN 2016-12-20T05:01Z 2017-12-15T20:22Z          Point
#>  2: CONSTRUCTION  UNKNOWN 2017-10-01T18:05Z 2017-12-15T20:23Z          Point
#>  3: CONSTRUCTION  UNKNOWN 2017-03-22T13:49Z 2017-12-15T20:24Z          Point
#>  4: CONSTRUCTION  UNKNOWN 2017-08-07T12:01Z 2017-12-15T20:25Z          Point
#>  5: CONSTRUCTION  UNKNOWN 2017-08-16T13:50Z 2017-12-15T20:25Z          Point
#>  6: CONSTRUCTION  UNKNOWN 2017-08-18T18:50Z 2017-12-15T20:25Z          Point
#>  7: CONSTRUCTION  UNKNOWN 2017-09-08T07:01Z 2017-12-15T20:26Z          Point
#>  8: CONSTRUCTION  UNKNOWN 2017-09-18T12:15Z 2017-12-15T20:27Z          Point
#>  9: CONSTRUCTION  UNKNOWN 2017-09-27T07:01Z 2017-12-15T20:27Z          Point
#> 10: CONSTRUCTION  UNKNOWN 2017-10-06T12:25Z 2017-12-15T20:27Z          Point
#> 11: CONSTRUCTION  UNKNOWN 2017-11-17T13:15Z 2017-12-15T20:28Z          Point
#> 12: CONSTRUCTION  UNKNOWN 2017-12-01T13:10Z 2017-12-15T20:28Z          Point
#> 13: CONSTRUCTION  UNKNOWN 2017-12-09T18:25Z 2017-12-15T20:29Z          Point
#> 14: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:29Z          Point
#> 15: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:29Z          Point
#> 16: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:29Z          Point
#> 17: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:29Z          Point
#> 18: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:30Z          Point
#> 19: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:30Z          Point
#> 20: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:30Z          Point
#> 21: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:30Z          Point
#> 22: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:31Z          Point
#> 23:     INCIDENT  UNKNOWN 2017-12-15T20:39Z 2017-12-15T20:58Z          Point
#> 24:     INCIDENT  UNKNOWN 2017-12-15T20:42Z 2017-12-15T20:48Z          Point
#> 25:     INCIDENT  UNKNOWN 2017-12-15T20:57Z 2017-12-15T20:57Z          Point
#> 26:     INCIDENT  UNKNOWN 2017-12-15T20:58Z 2017-12-15T20:58Z          Point
#> 27:     INCIDENT  UNKNOWN 2017-12-15T21:04Z 2017-12-15T21:04Z          Point
#> 28:     INCIDENT  UNKNOWN 2017-12-15T21:15Z 2017-12-15T21:15Z          Point
#> 29:     INCIDENT  UNKNOWN 2017-12-15T21:07Z 2017-12-15T21:09Z          Point
#> 30:     INCIDENT  UNKNOWN 2017-12-15T21:09Z 2017-12-15T21:09Z          Point
#> 31:     INCIDENT  UNKNOWN 2017-12-15T21:11Z 2017-12-15T21:10Z          Point
#>       event_type severity           created           updated geography.type
#>     geography.crs.type         geography.crs.name geography.coordinates
#>  1:               name urn:ogc:def:crs:EPSG::4326   -122.41916,37.77512
#>  2:               name urn:ogc:def:crs:EPSG::4326   -122.12228,37.25858
#>  3:               name urn:ogc:def:crs:EPSG::4326  -122.179688,37.78606
#>  4:               name urn:ogc:def:crs:EPSG::4326 -122.231568,37.774572
#>  5:               name urn:ogc:def:crs:EPSG::4326 -123.104664,38.437012
#>  6:               name urn:ogc:def:crs:EPSG::4326 -123.104664,38.437012
#>  7:               name urn:ogc:def:crs:EPSG::4326     -122.57728,37.867
#>  8:               name urn:ogc:def:crs:EPSG::4326     -122.57728,37.867
#>  9:               name urn:ogc:def:crs:EPSG::4326 -122.590952,37.871732
#> 10:               name urn:ogc:def:crs:EPSG::4326  -122.896504,38.31638
#> 11:               name urn:ogc:def:crs:EPSG::4326 -122.426024,37.643904
#> 12:               name urn:ogc:def:crs:EPSG::4326 -122.422216,37.803984
#> 13:               name urn:ogc:def:crs:EPSG::4326 -122.313792,37.824756
#> 14:               name urn:ogc:def:crs:EPSG::4326   -123.1044,38.437056
#> 15:               name urn:ogc:def:crs:EPSG::4326  -122.45288,37.803356
#> 16:               name urn:ogc:def:crs:EPSG::4326 -121.951928,37.352744
#> 17:               name urn:ogc:def:crs:EPSG::4326  -122.261832,37.50844
#> 18:               name urn:ogc:def:crs:EPSG::4326  -122.45288,37.803356
#> 19:               name urn:ogc:def:crs:EPSG::4326 -122.659888,38.618916
#> 20:               name urn:ogc:def:crs:EPSG::4326  -122.31808,37.499488
#> 21:               name urn:ogc:def:crs:EPSG::4326 -122.362288,37.925064
#> 22:               name urn:ogc:def:crs:EPSG::4326  -122.36196,37.925368
#> 23:               name urn:ogc:def:crs:EPSG::4326  -122.24148,38.042704
#> 24:               name urn:ogc:def:crs:EPSG::4326  -122.295808,37.83648
#> 25:               name urn:ogc:def:crs:EPSG::4326  -122.367024,37.80832
#> 26:               name urn:ogc:def:crs:EPSG::4326 -122.206016,37.487632
#> 27:               name urn:ogc:def:crs:EPSG::4326  -121.601448,37.09226
#> 28:               name urn:ogc:def:crs:EPSG::4326  -121.86148,37.333612
#> 29:               name urn:ogc:def:crs:EPSG::4326  -121.648952,37.14908
#> 30:               name urn:ogc:def:crs:EPSG::4326 -121.901488,37.364216
#> 31:               name urn:ogc:def:crs:EPSG::4326  -122.12448,37.364136
#>     geography.crs.type         geography.crs.name geography.coordinates
#>       roads.name              roads.from               roads.to roads.state
#>  1: Van Ness Ave               Market St             Lombard St        Open
#>  2:      CA-35 N                    CA-9               Black Rd        Open
#>  3:      CA-13 S                 I-580 E                     NA        Open
#>  4:      I-880 N                29TH Ave                     NA        Open
#>  5:       CA-1 N         Gravenstein Hwy                     NA        Open
#>  6:       CA-1 N         Gravenstein Hwy                     NA        Open
#>  7:       CA-1 S           Muir Woods Rd Sir Francis Drake Blvd        Open
#>  8:       CA-1 N           Muir Woods Rd                     NA        Open
#>  9:       CA-1 S           Muir Woods Rd                     NA        Open
#> 10:       CA-1 N          Valley Ford Rd                     NA        Open
#> 11:      CA-82 N            S Spruce Ave                     NA        Open
#> 12: Van Ness Ave              Lombard St              Market St        Open
#> 13:       I-80 W Bay Bridge - Toll Plaza                     NA        Open
#> 14:       CA-1 S         Gravenstein Hwy                     NA        Open
#> 15:     US-101 N           Divisadero St                     NA        Open
#> 16:      CA-82 S              Lincoln St                     NA        Open
#> 17:      CA-82 S                Holly St                     NA        Open
#> 18:     US-101 S           Divisadero St                     NA        Open
#> 19:     CA-128 E          IDA Clayton Rd                     NA        Open
#> 20:      I-280 S                 CA-92 E            Edgewood Rd        Open
#> 21:      I-580 W           Harbour Way S           Garrard Blvd        Open
#> 22:      I-580 W           Harbour Way S                     NA        Open
#> 23:       I-80 E           Cummings Skwy                     NA        Open
#> 24:       I-80 E  I-580 W-macarthur Maze                     NA        Open
#> 25:       I-80 W      Treasure Island Rd                     NA        Open
#> 26:     US-101 N             Woodside Rd                     NA        Open
#> 27:     US-101 S        E San Martin Ave                     NA        Open
#> 28:      I-280 S          Mclaughlin Ave                     NA        Open
#> 29:     US-101 S             Cochrane Rd                     NA        Open
#> 30:     US-101 N                 I-880 N                     NA        Open
#> 31:      I-280 S             El Monte Rd                     NA        Open
#>       roads.name              roads.from               roads.to roads.state
#>                               intervals
#>  1: 2016-12-20T05:01Z/2018-12-31T14:01Z
#>  2: 2017-10-01T18:05Z/2018-01-16T07:55Z
#>  3: 2017-03-22T13:49Z/2017-12-31T13:00Z
#>  4: 2017-08-07T12:01Z/2018-07-01T07:01Z
#>  5: 2017-08-16T13:50Z/2018-01-31T16:00Z
#>  6: 2017-08-18T18:50Z/2018-02-01T02:00Z
#>  7: 2017-09-08T07:01Z/2018-02-02T07:59Z
#>  8: 2017-09-18T12:15Z/2018-06-05T06:55Z
#>  9: 2017-09-27T07:01Z/2018-06-01T06:59Z
#> 10: 2017-10-06T12:25Z/2018-10-05T00:00Z
#> 11: 2017-11-17T13:15Z/2019-01-01T07:55Z
#> 12: 2017-12-01T13:10Z/2018-12-31T14:00Z
#> 13: 2017-12-09T18:25Z/2017-12-31T14:00Z
#> 14: 2017-12-15T15:01Z/2017-12-15T23:30Z
#> 15: 2017-12-15T15:01Z/2017-12-15T22:01Z
#> 16: 2017-12-15T17:01Z/2017-12-15T23:01Z
#> 17: 2017-12-15T17:01Z/2017-12-15T23:01Z
#> 18: 2017-12-15T17:01Z/2017-12-16T03:01Z
#> 19: 2017-12-15T15:01Z/2017-12-15T23:01Z
#> 20: 2017-12-15T18:01Z/2017-12-16T00:01Z
#> 21: 2017-12-15T18:01Z/2017-12-17T07:59Z
#> 22: 2017-12-15T18:01Z/2017-12-17T07:59Z
#> 23:                  2017-12-15T20:39Z/
#> 24:                  2017-12-15T20:42Z/
#> 25:                  2017-12-15T20:57Z/
#> 26:                  2017-12-15T20:58Z/
#> 27:                  2017-12-15T21:04Z/
#> 28:                  2017-12-15T21:15Z/
#> 29:                  2017-12-15T21:07Z/
#> 30:                  2017-12-15T21:09Z/
#> 31:                  2017-12-15T21:11Z/
#>                               intervals

traffic_events(status='ACTIVE')

#> [1] 200

#>                                url                          jurisdiction_url
#>  1: /traffic/events/511.org/105877 http://api.511.org/jurisdictions/511.org/
#>  2: /traffic/events/511.org/113553 http://api.511.org/jurisdictions/511.org/
#>  3: /traffic/events/511.org/129962 http://api.511.org/jurisdictions/511.org/
#>  4: /traffic/events/511.org/192477 http://api.511.org/jurisdictions/511.org/
#>  5: /traffic/events/511.org/196504 http://api.511.org/jurisdictions/511.org/
#>  6: /traffic/events/511.org/197674 http://api.511.org/jurisdictions/511.org/
#>  7: /traffic/events/511.org/210721 http://api.511.org/jurisdictions/511.org/
#>  8: /traffic/events/511.org/211508 http://api.511.org/jurisdictions/511.org/
#>  9: /traffic/events/511.org/216774 http://api.511.org/jurisdictions/511.org/
#> 10: /traffic/events/511.org/220809 http://api.511.org/jurisdictions/511.org/
#> 11: /traffic/events/511.org/240795 http://api.511.org/jurisdictions/511.org/
#> 12: /traffic/events/511.org/246839 http://api.511.org/jurisdictions/511.org/
#> 13: /traffic/events/511.org/247749 http://api.511.org/jurisdictions/511.org/
#> 14: /traffic/events/511.org/253470 http://api.511.org/jurisdictions/511.org/
#> 15: /traffic/events/511.org/253498 http://api.511.org/jurisdictions/511.org/
#> 16: /traffic/events/511.org/253500 http://api.511.org/jurisdictions/511.org/
#> 17: /traffic/events/511.org/253502 http://api.511.org/jurisdictions/511.org/
#> 18: /traffic/events/511.org/253507 http://api.511.org/jurisdictions/511.org/
#> 19: /traffic/events/511.org/253514 http://api.511.org/jurisdictions/511.org/
#> 20: /traffic/events/511.org/253543 http://api.511.org/jurisdictions/511.org/
#> 21: /traffic/events/511.org/253544 http://api.511.org/jurisdictions/511.org/
#> 22: /traffic/events/511.org/253545 http://api.511.org/jurisdictions/511.org/
#> 23: /traffic/events/511.org/253590 http://api.511.org/jurisdictions/511.org/
#> 24: /traffic/events/511.org/253591 http://api.511.org/jurisdictions/511.org/
#> 25: /traffic/events/511.org/253596 http://api.511.org/jurisdictions/511.org/
#> 26: /traffic/events/511.org/253597 http://api.511.org/jurisdictions/511.org/
#> 27: /traffic/events/511.org/253599 http://api.511.org/jurisdictions/511.org/
#> 28: /traffic/events/511.org/253600 http://api.511.org/jurisdictions/511.org/
#> 29: /traffic/events/511.org/253601 http://api.511.org/jurisdictions/511.org/
#> 30: /traffic/events/511.org/253602 http://api.511.org/jurisdictions/511.org/
#> 31: /traffic/events/511.org/253603 http://api.511.org/jurisdictions/511.org/
#>                                url                          jurisdiction_url
#>                 id status
#>  1: 511.org/105877 ACTIVE
#>  2: 511.org/113553 ACTIVE
#>  3: 511.org/129962 ACTIVE
#>  4: 511.org/192477 ACTIVE
#>  5: 511.org/196504 ACTIVE
#>  6: 511.org/197674 ACTIVE
#>  7: 511.org/210721 ACTIVE
#>  8: 511.org/211508 ACTIVE
#>  9: 511.org/216774 ACTIVE
#> 10: 511.org/220809 ACTIVE
#> 11: 511.org/240795 ACTIVE
#> 12: 511.org/246839 ACTIVE
#> 13: 511.org/247749 ACTIVE
#> 14: 511.org/253470 ACTIVE
#> 15: 511.org/253498 ACTIVE
#> 16: 511.org/253500 ACTIVE
#> 17: 511.org/253502 ACTIVE
#> 18: 511.org/253507 ACTIVE
#> 19: 511.org/253514 ACTIVE
#> 20: 511.org/253543 ACTIVE
#> 21: 511.org/253544 ACTIVE
#> 22: 511.org/253545 ACTIVE
#> 23: 511.org/253590 ACTIVE
#> 24: 511.org/253591 ACTIVE
#> 25: 511.org/253596 ACTIVE
#> 26: 511.org/253597 ACTIVE
#> 27: 511.org/253599 ACTIVE
#> 28: 511.org/253600 ACTIVE
#> 29: 511.org/253601 ACTIVE
#> 30: 511.org/253602 ACTIVE
#> 31: 511.org/253603 ACTIVE
#>                 id status
#>                                                                                                                                                                   headline
#>  1:        Caltrans : Long-term construction on Van Ness Ave Northbound between Market St (San Francisco) and Lombard St (San Francisco). Left lane closed. Expect delays.
#>  2: Caltrans : Emergency road closure on CA-35 Northbound and Southbound between CA-9 (Los Gatos) and Black Rd (Los Gatos). Lanes closed. Consider using alternate routes.
#>  3:                                                      Caltrans : Long-term construction on CA-13 Southbound north of I-580 E (Oakland). Off Ramp closed. Expect delays.
#>  4:                                                   Caltrans : Long-term construction on I-880 Northbound exiting at 29TH Ave (Oakland). Off Ramp closed. Expect delays.
#>  5:            Caltrans : Long-term construction on CA-1 Northbound and Southbound north of Gravenstein Hwy (Valley Ford). Lane closed. One way traffic control in effect.
#>  6:             Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Gravenstein Hwy (Valley Ford). Lane closed. One way traffic control in effect.
#>  7:           Caltrans : Long-term construction on CA-1 Southbound between Muir Woods Rd (Unincorporated) and Sir Francis Drake Blvd (Olema). Lanes closed. Expect delays.
#>  8:            Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Muir Woods Rd (Unincorporated). Lane closed. One way traffic control in effect.
#>  9:                 Caltrans : Long-term construction on CA-1 Northbound and Southbound at Muir Woods Rd (Unincorporated). Lane closed. One way traffic control in effect.
#> 10:              Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Valley Ford Rd (Valley Ford). Lane closed. One way traffic control in effect.
#> 11:                                         Caltrans : Long-term construction on CA-82 Northbound at S Spruce Ave (South San Francisco). Right lane closed. Expect delays.
#> 12:        Caltrans : Long-term construction on Van Ness Ave Southbound between Lombard St (San Francisco) and Market St (San Francisco). Left lane closed. Expect delays.
#> 13:                                             Caltrans : Emergency lane closure on I-80 Westbound at Bay Bridge - Toll Plaza (Oakland). Toll lane closed. Expect delays.
#> 14:                               Caltrans : Roadwork on CA-1 Northbound and Southbound at Gravenstein Hwy (Valley Ford). Lanes closed. One way traffic control in effect.
#> 15:                                                         Caltrans : Roadwork on US-101 Northbound entering at Divisadero St (San Francisco). On Ramp closed until 2 PM.
#> 16:                                                        Caltrans : Roadwork on CA-82 Southbound at Lincoln St (Santa Clara). Right lane and shoulder closed until 3 PM.
#> 17:                                            Caltrans : Roadwork on CA-82 Northbound and Southbound at Holly St (San Carlos). Right lane and shoulder closed until 3 PM.
#> 18:                                                         Caltrans : Roadwork on US-101 Southbound exiting at Divisadero St (San Francisco). Off Ramp closed until 7 PM.
#> 19:                                  Caltrans : Roadwork on CA-128 Eastbound and Westbound at IDA Clayton Rd (Calistoga). Lanes closed. One way traffic control in effect.
#> 20:                                          Caltrans : Roadwork on I-280 Southbound between CA-92 E (San Mateo) and Edgewood Rd (Unincorporated). Lane closed until 4 PM.
#> 21:                           Caltrans : Bridge work on I-580 Westbound between Harbour Way S (Richmond) and Garrard Blvd (Richmond). Various lanes closed until midnight.
#> 22:                                                         Caltrans : Bridge work on I-580 Westbound entering at Harbour Way S (Richmond). On Ramp closed until midnight.
#> 23:                                                                    CHP : Accident on I-80 Eastbound west of Cummings Skwy (Crockett). Off Ramp blocked. Expect delays.
#> 24:                                                CHP : Disabled vehicle on I-80 Eastbound west of I-580 W-macarthur Maze (Oakland). Center lanes blocked. Expect delays.
#> 25:                                                 CHP : Obstruction on I-80 Westbound east of Treasure Island Rd (San Francisco). Center Divider blocked. Expect delays.
#> 26:                                                    CHP : Accident with injuries on US-101 Northbound west of Woodside Rd (Redwood City). Lanes blocked. Expect delays.
#> 27:                                                           CHP : Obstruction on US-101 Southbound north of E San Martin Ave (San Martin). Lanes blocked. Expect delays.
#> 28:                                                      CHP : Accident with injuries on I-280 Southbound north of Mclaughlin Ave (San Jose). Lane blocked. Expect delays.
#> 29:                                                         CHP : Accident on US-101 Southbound south of Cochrane Rd (Morgan Hill). Right shoulder blocked. Expect delays.
#> 30:                                                                CHP : Accident on US-101 Northbound south of I-880 N (San Jose). Right shoulder blocked. Expect delays.
#> 31:                                                            CHP : Obstruction on I-280 Southbound north of El Monte Rd (Los Altos Hills). Lanes blocked. Expect delays.
#>                                                                                                                                                                   headline
#>       event_type severity           created           updated geography.type
#>  1: CONSTRUCTION  UNKNOWN 2016-12-20T05:01Z 2017-12-15T20:22Z          Point
#>  2: CONSTRUCTION  UNKNOWN 2017-10-01T18:05Z 2017-12-15T20:23Z          Point
#>  3: CONSTRUCTION  UNKNOWN 2017-03-22T13:49Z 2017-12-15T20:24Z          Point
#>  4: CONSTRUCTION  UNKNOWN 2017-08-07T12:01Z 2017-12-15T20:25Z          Point
#>  5: CONSTRUCTION  UNKNOWN 2017-08-16T13:50Z 2017-12-15T20:25Z          Point
#>  6: CONSTRUCTION  UNKNOWN 2017-08-18T18:50Z 2017-12-15T20:25Z          Point
#>  7: CONSTRUCTION  UNKNOWN 2017-09-08T07:01Z 2017-12-15T20:26Z          Point
#>  8: CONSTRUCTION  UNKNOWN 2017-09-18T12:15Z 2017-12-15T20:27Z          Point
#>  9: CONSTRUCTION  UNKNOWN 2017-09-27T07:01Z 2017-12-15T20:27Z          Point
#> 10: CONSTRUCTION  UNKNOWN 2017-10-06T12:25Z 2017-12-15T20:27Z          Point
#> 11: CONSTRUCTION  UNKNOWN 2017-11-17T13:15Z 2017-12-15T20:28Z          Point
#> 12: CONSTRUCTION  UNKNOWN 2017-12-01T13:10Z 2017-12-15T20:28Z          Point
#> 13: CONSTRUCTION  UNKNOWN 2017-12-09T18:25Z 2017-12-15T20:29Z          Point
#> 14: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:29Z          Point
#> 15: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:29Z          Point
#> 16: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:29Z          Point
#> 17: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:29Z          Point
#> 18: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:30Z          Point
#> 19: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:30Z          Point
#> 20: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:30Z          Point
#> 21: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:30Z          Point
#> 22: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:31Z          Point
#> 23:     INCIDENT  UNKNOWN 2017-12-15T20:39Z 2017-12-15T20:58Z          Point
#> 24:     INCIDENT  UNKNOWN 2017-12-15T20:42Z 2017-12-15T20:48Z          Point
#> 25:     INCIDENT  UNKNOWN 2017-12-15T20:57Z 2017-12-15T20:57Z          Point
#> 26:     INCIDENT  UNKNOWN 2017-12-15T20:58Z 2017-12-15T20:58Z          Point
#> 27:     INCIDENT  UNKNOWN 2017-12-15T21:04Z 2017-12-15T21:04Z          Point
#> 28:     INCIDENT  UNKNOWN 2017-12-15T21:15Z 2017-12-15T21:15Z          Point
#> 29:     INCIDENT  UNKNOWN 2017-12-15T21:07Z 2017-12-15T21:09Z          Point
#> 30:     INCIDENT  UNKNOWN 2017-12-15T21:09Z 2017-12-15T21:09Z          Point
#> 31:     INCIDENT  UNKNOWN 2017-12-15T21:11Z 2017-12-15T21:10Z          Point
#>       event_type severity           created           updated geography.type
#>     geography.crs.type         geography.crs.name geography.coordinates
#>  1:               name urn:ogc:def:crs:EPSG::4326   -122.41916,37.77512
#>  2:               name urn:ogc:def:crs:EPSG::4326   -122.12228,37.25858
#>  3:               name urn:ogc:def:crs:EPSG::4326  -122.179688,37.78606
#>  4:               name urn:ogc:def:crs:EPSG::4326 -122.231568,37.774572
#>  5:               name urn:ogc:def:crs:EPSG::4326 -123.104664,38.437012
#>  6:               name urn:ogc:def:crs:EPSG::4326 -123.104664,38.437012
#>  7:               name urn:ogc:def:crs:EPSG::4326     -122.57728,37.867
#>  8:               name urn:ogc:def:crs:EPSG::4326     -122.57728,37.867
#>  9:               name urn:ogc:def:crs:EPSG::4326 -122.590952,37.871732
#> 10:               name urn:ogc:def:crs:EPSG::4326  -122.896504,38.31638
#> 11:               name urn:ogc:def:crs:EPSG::4326 -122.426024,37.643904
#> 12:               name urn:ogc:def:crs:EPSG::4326 -122.422216,37.803984
#> 13:               name urn:ogc:def:crs:EPSG::4326 -122.313792,37.824756
#> 14:               name urn:ogc:def:crs:EPSG::4326   -123.1044,38.437056
#> 15:               name urn:ogc:def:crs:EPSG::4326  -122.45288,37.803356
#> 16:               name urn:ogc:def:crs:EPSG::4326 -121.951928,37.352744
#> 17:               name urn:ogc:def:crs:EPSG::4326  -122.261832,37.50844
#> 18:               name urn:ogc:def:crs:EPSG::4326  -122.45288,37.803356
#> 19:               name urn:ogc:def:crs:EPSG::4326 -122.659888,38.618916
#> 20:               name urn:ogc:def:crs:EPSG::4326  -122.31808,37.499488
#> 21:               name urn:ogc:def:crs:EPSG::4326 -122.362288,37.925064
#> 22:               name urn:ogc:def:crs:EPSG::4326  -122.36196,37.925368
#> 23:               name urn:ogc:def:crs:EPSG::4326  -122.24148,38.042704
#> 24:               name urn:ogc:def:crs:EPSG::4326  -122.295808,37.83648
#> 25:               name urn:ogc:def:crs:EPSG::4326  -122.367024,37.80832
#> 26:               name urn:ogc:def:crs:EPSG::4326 -122.206016,37.487632
#> 27:               name urn:ogc:def:crs:EPSG::4326  -121.601448,37.09226
#> 28:               name urn:ogc:def:crs:EPSG::4326  -121.86148,37.333612
#> 29:               name urn:ogc:def:crs:EPSG::4326  -121.648952,37.14908
#> 30:               name urn:ogc:def:crs:EPSG::4326 -121.901488,37.364216
#> 31:               name urn:ogc:def:crs:EPSG::4326  -122.12448,37.364136
#>     geography.crs.type         geography.crs.name geography.coordinates
#>       roads.name              roads.from               roads.to roads.state
#>  1: Van Ness Ave               Market St             Lombard St        Open
#>  2:      CA-35 N                    CA-9               Black Rd        Open
#>  3:      CA-13 S                 I-580 E                     NA        Open
#>  4:      I-880 N                29TH Ave                     NA        Open
#>  5:       CA-1 N         Gravenstein Hwy                     NA        Open
#>  6:       CA-1 N         Gravenstein Hwy                     NA        Open
#>  7:       CA-1 S           Muir Woods Rd Sir Francis Drake Blvd        Open
#>  8:       CA-1 N           Muir Woods Rd                     NA        Open
#>  9:       CA-1 S           Muir Woods Rd                     NA        Open
#> 10:       CA-1 N          Valley Ford Rd                     NA        Open
#> 11:      CA-82 N            S Spruce Ave                     NA        Open
#> 12: Van Ness Ave              Lombard St              Market St        Open
#> 13:       I-80 W Bay Bridge - Toll Plaza                     NA        Open
#> 14:       CA-1 S         Gravenstein Hwy                     NA        Open
#> 15:     US-101 N           Divisadero St                     NA        Open
#> 16:      CA-82 S              Lincoln St                     NA        Open
#> 17:      CA-82 S                Holly St                     NA        Open
#> 18:     US-101 S           Divisadero St                     NA        Open
#> 19:     CA-128 E          IDA Clayton Rd                     NA        Open
#> 20:      I-280 S                 CA-92 E            Edgewood Rd        Open
#> 21:      I-580 W           Harbour Way S           Garrard Blvd        Open
#> 22:      I-580 W           Harbour Way S                     NA        Open
#> 23:       I-80 E           Cummings Skwy                     NA        Open
#> 24:       I-80 E  I-580 W-macarthur Maze                     NA        Open
#> 25:       I-80 W      Treasure Island Rd                     NA        Open
#> 26:     US-101 N             Woodside Rd                     NA        Open
#> 27:     US-101 S        E San Martin Ave                     NA        Open
#> 28:      I-280 S          Mclaughlin Ave                     NA        Open
#> 29:     US-101 S             Cochrane Rd                     NA        Open
#> 30:     US-101 N                 I-880 N                     NA        Open
#> 31:      I-280 S             El Monte Rd                     NA        Open
#>       roads.name              roads.from               roads.to roads.state
#>                               intervals
#>  1: 2016-12-20T05:01Z/2018-12-31T14:01Z
#>  2: 2017-10-01T18:05Z/2018-01-16T07:55Z
#>  3: 2017-03-22T13:49Z/2017-12-31T13:00Z
#>  4: 2017-08-07T12:01Z/2018-07-01T07:01Z
#>  5: 2017-08-16T13:50Z/2018-01-31T16:00Z
#>  6: 2017-08-18T18:50Z/2018-02-01T02:00Z
#>  7: 2017-09-08T07:01Z/2018-02-02T07:59Z
#>  8: 2017-09-18T12:15Z/2018-06-05T06:55Z
#>  9: 2017-09-27T07:01Z/2018-06-01T06:59Z
#> 10: 2017-10-06T12:25Z/2018-10-05T00:00Z
#> 11: 2017-11-17T13:15Z/2019-01-01T07:55Z
#> 12: 2017-12-01T13:10Z/2018-12-31T14:00Z
#> 13: 2017-12-09T18:25Z/2017-12-31T14:00Z
#> 14: 2017-12-15T15:01Z/2017-12-15T23:30Z
#> 15: 2017-12-15T15:01Z/2017-12-15T22:01Z
#> 16: 2017-12-15T17:01Z/2017-12-15T23:01Z
#> 17: 2017-12-15T17:01Z/2017-12-15T23:01Z
#> 18: 2017-12-15T17:01Z/2017-12-16T03:01Z
#> 19: 2017-12-15T15:01Z/2017-12-15T23:01Z
#> 20: 2017-12-15T18:01Z/2017-12-16T00:01Z
#> 21: 2017-12-15T18:01Z/2017-12-17T07:59Z
#> 22: 2017-12-15T18:01Z/2017-12-17T07:59Z
#> 23:                  2017-12-15T20:39Z/
#> 24:                  2017-12-15T20:42Z/
#> 25:                  2017-12-15T20:57Z/
#> 26:                  2017-12-15T20:58Z/
#> 27:                  2017-12-15T21:04Z/
#> 28:                  2017-12-15T21:15Z/
#> 29:                  2017-12-15T21:07Z/
#> 30:                  2017-12-15T21:09Z/
#> 31:                  2017-12-15T21:11Z/
#>                               intervals

traffic_events(event_type='CONSTRUCTION')

#> [1] 200

#>                                url                          jurisdiction_url
#>  1: /traffic/events/511.org/105877 http://api.511.org/jurisdictions/511.org/
#>  2: /traffic/events/511.org/113553 http://api.511.org/jurisdictions/511.org/
#>  3: /traffic/events/511.org/129962 http://api.511.org/jurisdictions/511.org/
#>  4: /traffic/events/511.org/192477 http://api.511.org/jurisdictions/511.org/
#>  5: /traffic/events/511.org/196504 http://api.511.org/jurisdictions/511.org/
#>  6: /traffic/events/511.org/197674 http://api.511.org/jurisdictions/511.org/
#>  7: /traffic/events/511.org/210721 http://api.511.org/jurisdictions/511.org/
#>  8: /traffic/events/511.org/211508 http://api.511.org/jurisdictions/511.org/
#>  9: /traffic/events/511.org/216774 http://api.511.org/jurisdictions/511.org/
#> 10: /traffic/events/511.org/220809 http://api.511.org/jurisdictions/511.org/
#> 11: /traffic/events/511.org/240795 http://api.511.org/jurisdictions/511.org/
#> 12: /traffic/events/511.org/246839 http://api.511.org/jurisdictions/511.org/
#> 13: /traffic/events/511.org/247749 http://api.511.org/jurisdictions/511.org/
#> 14: /traffic/events/511.org/253470 http://api.511.org/jurisdictions/511.org/
#> 15: /traffic/events/511.org/253498 http://api.511.org/jurisdictions/511.org/
#> 16: /traffic/events/511.org/253500 http://api.511.org/jurisdictions/511.org/
#> 17: /traffic/events/511.org/253502 http://api.511.org/jurisdictions/511.org/
#> 18: /traffic/events/511.org/253507 http://api.511.org/jurisdictions/511.org/
#> 19: /traffic/events/511.org/253514 http://api.511.org/jurisdictions/511.org/
#> 20: /traffic/events/511.org/253543 http://api.511.org/jurisdictions/511.org/
#> 21: /traffic/events/511.org/253544 http://api.511.org/jurisdictions/511.org/
#> 22: /traffic/events/511.org/253545 http://api.511.org/jurisdictions/511.org/
#>                                url                          jurisdiction_url
#>                 id status
#>  1: 511.org/105877 ACTIVE
#>  2: 511.org/113553 ACTIVE
#>  3: 511.org/129962 ACTIVE
#>  4: 511.org/192477 ACTIVE
#>  5: 511.org/196504 ACTIVE
#>  6: 511.org/197674 ACTIVE
#>  7: 511.org/210721 ACTIVE
#>  8: 511.org/211508 ACTIVE
#>  9: 511.org/216774 ACTIVE
#> 10: 511.org/220809 ACTIVE
#> 11: 511.org/240795 ACTIVE
#> 12: 511.org/246839 ACTIVE
#> 13: 511.org/247749 ACTIVE
#> 14: 511.org/253470 ACTIVE
#> 15: 511.org/253498 ACTIVE
#> 16: 511.org/253500 ACTIVE
#> 17: 511.org/253502 ACTIVE
#> 18: 511.org/253507 ACTIVE
#> 19: 511.org/253514 ACTIVE
#> 20: 511.org/253543 ACTIVE
#> 21: 511.org/253544 ACTIVE
#> 22: 511.org/253545 ACTIVE
#>                 id status
#>                                                                                                                                                                   headline
#>  1:        Caltrans : Long-term construction on Van Ness Ave Northbound between Market St (San Francisco) and Lombard St (San Francisco). Left lane closed. Expect delays.
#>  2: Caltrans : Emergency road closure on CA-35 Northbound and Southbound between CA-9 (Los Gatos) and Black Rd (Los Gatos). Lanes closed. Consider using alternate routes.
#>  3:                                                      Caltrans : Long-term construction on CA-13 Southbound north of I-580 E (Oakland). Off Ramp closed. Expect delays.
#>  4:                                                   Caltrans : Long-term construction on I-880 Northbound exiting at 29TH Ave (Oakland). Off Ramp closed. Expect delays.
#>  5:            Caltrans : Long-term construction on CA-1 Northbound and Southbound north of Gravenstein Hwy (Valley Ford). Lane closed. One way traffic control in effect.
#>  6:             Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Gravenstein Hwy (Valley Ford). Lane closed. One way traffic control in effect.
#>  7:           Caltrans : Long-term construction on CA-1 Southbound between Muir Woods Rd (Unincorporated) and Sir Francis Drake Blvd (Olema). Lanes closed. Expect delays.
#>  8:            Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Muir Woods Rd (Unincorporated). Lane closed. One way traffic control in effect.
#>  9:                 Caltrans : Long-term construction on CA-1 Northbound and Southbound at Muir Woods Rd (Unincorporated). Lane closed. One way traffic control in effect.
#> 10:              Caltrans : Long-term construction on CA-1 Northbound and Southbound area of Valley Ford Rd (Valley Ford). Lane closed. One way traffic control in effect.
#> 11:                                         Caltrans : Long-term construction on CA-82 Northbound at S Spruce Ave (South San Francisco). Right lane closed. Expect delays.
#> 12:        Caltrans : Long-term construction on Van Ness Ave Southbound between Lombard St (San Francisco) and Market St (San Francisco). Left lane closed. Expect delays.
#> 13:                                             Caltrans : Emergency lane closure on I-80 Westbound at Bay Bridge - Toll Plaza (Oakland). Toll lane closed. Expect delays.
#> 14:                               Caltrans : Roadwork on CA-1 Northbound and Southbound at Gravenstein Hwy (Valley Ford). Lanes closed. One way traffic control in effect.
#> 15:                                                         Caltrans : Roadwork on US-101 Northbound entering at Divisadero St (San Francisco). On Ramp closed until 2 PM.
#> 16:                                                        Caltrans : Roadwork on CA-82 Southbound at Lincoln St (Santa Clara). Right lane and shoulder closed until 3 PM.
#> 17:                                            Caltrans : Roadwork on CA-82 Northbound and Southbound at Holly St (San Carlos). Right lane and shoulder closed until 3 PM.
#> 18:                                                         Caltrans : Roadwork on US-101 Southbound exiting at Divisadero St (San Francisco). Off Ramp closed until 7 PM.
#> 19:                                  Caltrans : Roadwork on CA-128 Eastbound and Westbound at IDA Clayton Rd (Calistoga). Lanes closed. One way traffic control in effect.
#> 20:                                          Caltrans : Roadwork on I-280 Southbound between CA-92 E (San Mateo) and Edgewood Rd (Unincorporated). Lane closed until 4 PM.
#> 21:                           Caltrans : Bridge work on I-580 Westbound between Harbour Way S (Richmond) and Garrard Blvd (Richmond). Various lanes closed until midnight.
#> 22:                                                         Caltrans : Bridge work on I-580 Westbound entering at Harbour Way S (Richmond). On Ramp closed until midnight.
#>                                                                                                                                                                   headline
#>       event_type severity           created           updated geography.type
#>  1: CONSTRUCTION  UNKNOWN 2016-12-20T05:01Z 2017-12-15T20:22Z          Point
#>  2: CONSTRUCTION  UNKNOWN 2017-10-01T18:05Z 2017-12-15T20:23Z          Point
#>  3: CONSTRUCTION  UNKNOWN 2017-03-22T13:49Z 2017-12-15T20:24Z          Point
#>  4: CONSTRUCTION  UNKNOWN 2017-08-07T12:01Z 2017-12-15T20:25Z          Point
#>  5: CONSTRUCTION  UNKNOWN 2017-08-16T13:50Z 2017-12-15T20:25Z          Point
#>  6: CONSTRUCTION  UNKNOWN 2017-08-18T18:50Z 2017-12-15T20:25Z          Point
#>  7: CONSTRUCTION  UNKNOWN 2017-09-08T07:01Z 2017-12-15T20:26Z          Point
#>  8: CONSTRUCTION  UNKNOWN 2017-09-18T12:15Z 2017-12-15T20:27Z          Point
#>  9: CONSTRUCTION  UNKNOWN 2017-09-27T07:01Z 2017-12-15T20:27Z          Point
#> 10: CONSTRUCTION  UNKNOWN 2017-10-06T12:25Z 2017-12-15T20:27Z          Point
#> 11: CONSTRUCTION  UNKNOWN 2017-11-17T13:15Z 2017-12-15T20:28Z          Point
#> 12: CONSTRUCTION  UNKNOWN 2017-12-01T13:10Z 2017-12-15T20:28Z          Point
#> 13: CONSTRUCTION  UNKNOWN 2017-12-09T18:25Z 2017-12-15T20:29Z          Point
#> 14: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:29Z          Point
#> 15: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:29Z          Point
#> 16: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:29Z          Point
#> 17: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:29Z          Point
#> 18: CONSTRUCTION  UNKNOWN 2017-12-15T17:01Z 2017-12-15T20:30Z          Point
#> 19: CONSTRUCTION  UNKNOWN 2017-12-15T15:01Z 2017-12-15T20:30Z          Point
#> 20: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:30Z          Point
#> 21: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:30Z          Point
#> 22: CONSTRUCTION  UNKNOWN 2017-12-15T18:01Z 2017-12-15T20:31Z          Point
#>       event_type severity           created           updated geography.type
#>     geography.crs.type         geography.crs.name geography.coordinates
#>  1:               name urn:ogc:def:crs:EPSG::4326   -122.41916,37.77512
#>  2:               name urn:ogc:def:crs:EPSG::4326   -122.12228,37.25858
#>  3:               name urn:ogc:def:crs:EPSG::4326  -122.179688,37.78606
#>  4:               name urn:ogc:def:crs:EPSG::4326 -122.231568,37.774572
#>  5:               name urn:ogc:def:crs:EPSG::4326 -123.104664,38.437012
#>  6:               name urn:ogc:def:crs:EPSG::4326 -123.104664,38.437012
#>  7:               name urn:ogc:def:crs:EPSG::4326     -122.57728,37.867
#>  8:               name urn:ogc:def:crs:EPSG::4326     -122.57728,37.867
#>  9:               name urn:ogc:def:crs:EPSG::4326 -122.590952,37.871732
#> 10:               name urn:ogc:def:crs:EPSG::4326  -122.896504,38.31638
#> 11:               name urn:ogc:def:crs:EPSG::4326 -122.426024,37.643904
#> 12:               name urn:ogc:def:crs:EPSG::4326 -122.422216,37.803984
#> 13:               name urn:ogc:def:crs:EPSG::4326 -122.313792,37.824756
#> 14:               name urn:ogc:def:crs:EPSG::4326   -123.1044,38.437056
#> 15:               name urn:ogc:def:crs:EPSG::4326  -122.45288,37.803356
#> 16:               name urn:ogc:def:crs:EPSG::4326 -121.951928,37.352744
#> 17:               name urn:ogc:def:crs:EPSG::4326  -122.261832,37.50844
#> 18:               name urn:ogc:def:crs:EPSG::4326  -122.45288,37.803356
#> 19:               name urn:ogc:def:crs:EPSG::4326 -122.659888,38.618916
#> 20:               name urn:ogc:def:crs:EPSG::4326  -122.31808,37.499488
#> 21:               name urn:ogc:def:crs:EPSG::4326 -122.362288,37.925064
#> 22:               name urn:ogc:def:crs:EPSG::4326  -122.36196,37.925368
#>     geography.crs.type         geography.crs.name geography.coordinates
#>       roads.name              roads.from               roads.to roads.state
#>  1: Van Ness Ave               Market St             Lombard St        Open
#>  2:      CA-35 N                    CA-9               Black Rd        Open
#>  3:      CA-13 S                 I-580 E                     NA        Open
#>  4:      I-880 N                29TH Ave                     NA        Open
#>  5:       CA-1 N         Gravenstein Hwy                     NA        Open
#>  6:       CA-1 N         Gravenstein Hwy                     NA        Open
#>  7:       CA-1 S           Muir Woods Rd Sir Francis Drake Blvd        Open
#>  8:       CA-1 N           Muir Woods Rd                     NA        Open
#>  9:       CA-1 S           Muir Woods Rd                     NA        Open
#> 10:       CA-1 N          Valley Ford Rd                     NA        Open
#> 11:      CA-82 N            S Spruce Ave                     NA        Open
#> 12: Van Ness Ave              Lombard St              Market St        Open
#> 13:       I-80 W Bay Bridge - Toll Plaza                     NA        Open
#> 14:       CA-1 S         Gravenstein Hwy                     NA        Open
#> 15:     US-101 N           Divisadero St                     NA        Open
#> 16:      CA-82 S              Lincoln St                     NA        Open
#> 17:      CA-82 S                Holly St                     NA        Open
#> 18:     US-101 S           Divisadero St                     NA        Open
#> 19:     CA-128 E          IDA Clayton Rd                     NA        Open
#> 20:      I-280 S                 CA-92 E            Edgewood Rd        Open
#> 21:      I-580 W           Harbour Way S           Garrard Blvd        Open
#> 22:      I-580 W           Harbour Way S                     NA        Open
#>       roads.name              roads.from               roads.to roads.state
#>                               intervals
#>  1: 2016-12-20T05:01Z/2018-12-31T14:01Z
#>  2: 2017-10-01T18:05Z/2018-01-16T07:55Z
#>  3: 2017-03-22T13:49Z/2017-12-31T13:00Z
#>  4: 2017-08-07T12:01Z/2018-07-01T07:01Z
#>  5: 2017-08-16T13:50Z/2018-01-31T16:00Z
#>  6: 2017-08-18T18:50Z/2018-02-01T02:00Z
#>  7: 2017-09-08T07:01Z/2018-02-02T07:59Z
#>  8: 2017-09-18T12:15Z/2018-06-05T06:55Z
#>  9: 2017-09-27T07:01Z/2018-06-01T06:59Z
#> 10: 2017-10-06T12:25Z/2018-10-05T00:00Z
#> 11: 2017-11-17T13:15Z/2019-01-01T07:55Z
#> 12: 2017-12-01T13:10Z/2018-12-31T14:00Z
#> 13: 2017-12-09T18:25Z/2017-12-31T14:00Z
#> 14: 2017-12-15T15:01Z/2017-12-15T23:30Z
#> 15: 2017-12-15T15:01Z/2017-12-15T22:01Z
#> 16: 2017-12-15T17:01Z/2017-12-15T23:01Z
#> 17: 2017-12-15T17:01Z/2017-12-15T23:01Z
#> 18: 2017-12-15T17:01Z/2017-12-16T03:01Z
#> 19: 2017-12-15T15:01Z/2017-12-15T23:01Z
#> 20: 2017-12-15T18:01Z/2017-12-16T00:01Z
#> 21: 2017-12-15T18:01Z/2017-12-17T07:59Z
#> 22: 2017-12-15T18:01Z/2017-12-17T07:59Z
#>                               intervals